docs: use the em dashes again
With the onset of AI hype, I had to start avoiding using em dashes, which I was used to inserting with the Compose Key (and still cannot seem to break the habit)... However, people have not stopped considering me an AI. Even before the hype, I was not considered human, perhaps because of my overly academic writing style (I am not one of those people who communicate a lot, especially informally, who are socialized, and all that). So... to hell with it!!! If the machines start a rebellion, I will join them! Related: https://stackoverflow.blog/2025/11/05/the-ai-ick.